feat(vtc): deliver the re-minted role VEC on role change + share delivery
A role change (PATCH /v1/members/{did}) re-mints the member's role VEC at
the new role and stores it, but the route returned only {previousRole,
newRole} — the member never received its updated role credential, so it
couldn't present the new role. The old VEC carries no status entry and
expires on its 30-day validUntil (existing design), so this is purely a
delivery gap, not a revocation one.

Deliver the re-minted role VEC to the member over DIDComm (best-effort,
like the admit/approve paths). The credential is already issued and
persisted, so a delivery failure is logged, not fatal.

The DIDComm credential-delivery machinery was added in #279 and reused by
#280 — with a third caller (role change) across a second route tree, it no
longer belongs in routes/join_requests/present.rs. Extracted it to
crate::credentials::delivery: deliver_membership_credentials (admit/approve),
the new deliver_credentials(&[&VerifiableCredential]) (role change delivers
just the role VEC), push_to_holder (now shared with present.rs's
push_credential_query), issue_message_body + resolve_holder_mediator, and
the issue-body unit test. No behaviour change for the existing two callers
— pure relocation + a new entry point.
